Changed the BlockID to an enum class and uint8_t for memory efficiency

+++ src/blockRegistry.cpp
@@ -3,14 +3,14 @@
 
 
 bool BlockRegistry::is_solid(const BlockID& id){
-    return registry[id].solid;
+    return registry[static_cast<std::size_t>(id)].solid;
 }
 
 
 std::string BlockRegistry::getTextureName(const BlockID id, const FaceType faceType){
     switch(faceType){
-        case Side: return std::string(registry[id].textureNameSide);
-        case Top: return std::string(registry[id].textureNameTop);
-        default: return std::string(registry[id].textureNameBottom);
+        case Side: return std::string(registry[static_cast<std::size_t>(id)].textureNameSide);
+        case Top: return std::string(registry[static_cast<std::size_t>(id)].textureNameTop);
+        default: return std::string(registry[static_cast<std::size_t>(id)].textureNameBottom);
     }
 }
